requirements.txt
pinecone-client
transformers
PyPDF2
sentence-transformers
python-dotenv
ipykernel
groq
python-dotenv
langchain
langchain-groq
langchain_community
langgraph
pygraphviz
